+++ angular/src/app/data/rental.model.ts
@@ -1,11 +1,11 @@
+import { RentalUnit } from './rental-unit.model';
 /**
- * Represents the _Rental_ model
+ * Represents the _Rental_ modei
  *
  * ```yaml
  * id: string;
- * name: string;
- * occupancy: number;
- * type: string;
+ * lotNumber: string;
+ * unit: RentalUnit;
  * status: string;
  * price: number;
  * discountedPrice?: number;
@@ -13,11 +13,24 @@
  */
 export interface Rental {
   id: string;
-  name: string;
-  /** maximum number of people allowed */
-  occupancy: number;
-  /** type of site */
-  type: string;
+  lotNumber: string;
+  /** type of site
+   *  - plotSize
+   *    - an interface that is used to find out the size of a plot
+   *      ```yaml
+   *      width:number
+   *      height:number
+   *      ```
+   *  - amenities
+   *    - an interface that talks about what the campsite offers the camper
+   *      ```yaml
+   *      maxiumumCapacity:number
+   *      voltage:number
+   *      sewage:string
+   *      water:string
+   *      ```
+   */
+  unit: RentalUnit;
   /** booking status, one of:
    *  - available (neither booked nor currently in use)
    *  - booked (booked by someone else, but not in use)