+++ angular/src/app/modules/account/account-booking/account-booking.component.spec.ts
@@ -20,13 +20,15 @@ describe('AccountBookingComponent', () => {
     ],
     rentals: [
       {
-        id: 'string',
-        name: 'string',
-        status: 'string',
-        price: 0.0,
-        discountedPrice: 0.0,
-        type: 'string',
-        occupancy: 1,
+        id: '1',
+        lotNumber: '1',
+        unit: {
+          size: '5x5',
+          capacity: 2,
+          name: 'tent',
+        },
+        status: 'available',
+        price: 100,
       },
     ],
   };