+++ angular/src/environments/environment.stable.ts
@@ -1,6 +1,6 @@
 import { Environment } from './types/environment.type';
 import * as manifest from '../../package.json';
-import * as config from '../../angular.json';
+import * as ngConfig from '../../angular.json';
 
 export const environment: Environment = {
   config: 'app.config.json',
@@ -11,7 +11,8 @@ export const environment: Environment = {
     redirectUri: `${location.protocol}//${location.host}/oauth2/authorize`,
     scopes: ['openid', 'profile', 'email'],
   },
+  monitoring: 'https://ec61288c429544a6ad850b94de25004a@o388320.ingest.sentry.io/5229046',
   name: 'PROD',
   production: true,
-  release: `${config.defaultProject}@${manifest.version}`,
+  release: `${ngConfig.defaultProject}@${manifest.version}`,
 };